moved graph classification models to EGSteal folder
